requirements.txt
fastapi
PyMuPDF
pillow
pytesseract
python_docx
uvicorn[standard]
python-multipart
opencv-python-headless
easyocr
